split data cleaning functions from functions identifying invalid responses
